[level 2] Title: 이모티콘 할인행사, Time: 0.42 ms, Memory: 84.1 MB -BaekjoonHub

+++ 프로그래머스/2/150368. 이모티콘 할인행사/이모티콘 할인행사.java
@@ -18,7 +18,7 @@ class Solution {
             int subscribe = 0;
             double sum = 0;
             for(int[] user : users) {
-                double userDiscount = 1 - ((double) user[0] * 0.01);
+                double userDiscount = 1 - (user[0] * 0.01);
                 double tempSum = 0;
                 for(Emoticon emoticon : list) {
                     if(emoticon.discount > userDiscount) continue;
